# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stroopwafel Social API
  version: 1.2.0
  description: API-first social scheduling and publishing API for agents.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
  - apiKeyHeader: []
paths:
...
  /api/v1/posts:
    get:
      summary: List posts
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, scheduled, sent, failed]
        - in: query
          name: channel_id
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: scheduled_from
          schema: { type: string, format: date-time }
        - in: query
          name: scheduled_to
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Paginated posts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsListResponse'
        '400':
          $ref: '#/components/responses/APIError'
    post:
      summary: Create post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMutationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostMutationResponse'
        '400':
          $ref: '#/components/responses/APIError'
        '500':
          $ref: '#/components/responses/APIError'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    APIError:
      description: API error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
    Pagination:
      type: object
      required: [limit, offset, total, has_next, has_prev]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
        has_next: { type: boolean }
        has_prev: { type: boolean }
    APIError:
      type: object
      required: [error, error_code]
      properties:
        error:
          type: string
        error_code:
          type: string
    Post:
      type: object
      required: [id, text, status, created_at, updated_at, fail_count]
      properties:
        id: { type: integer }
        scheduled_at: { type: string, format: date-time, nullable: true }
        text: { type: string }
        status: { type: string, enum: [draft, scheduled, sent, failed] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        sent_at: { type: string, format: date-time, nullable: true }
        fail_count: { type: integer }
        last_error: { type: string, nullable: true }
        media_type: { type: string, nullable: true }
        media_url: { type: string, format: uri, nullable: true }
        next_retry_at: { type: string, format: date-time, nullable: true }
        channel_ids:
          type: array
          items: { type: integer }
    PostsListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PostMutationRequest:
      type: object
      required: [text, status]
      properties:
        scheduled_at: { type: string, format: date-time, nullable: true }
        text: { type: string }
        status: { type: string, enum: [draft, scheduled] }
        media_type: { type: string, enum: [link, image, video], nullable: true }
        media_url: { type: string, format: uri, nullable: true }
        channel_ids:
          type: array
          items: { type: integer }
    PostMutationResponse:
      type: object
      required: [post]
      properties:
        post:
          $ref: '#/components/schemas/Post'
        warnings:
          type: array
          items:
            type: object
            additionalProperties: true
...